requirements.txt
Pillow >= 10